requirements.txt
Flask
Flask-MySQL
Flask-SQLAlchemy
tensorflow
numpy
pandas
scikit-learn
python-dotenv
